+++ test/marskal_test.rb
@@ -1,11 +1,28 @@
 require 'test_helper'
 
-class MarskalTest < Minitest::Test
-  def test_that_it_has_a_version_number
-    refute_nil ::Marskal::VERSION
+describe 'Marskal' do
+  #---------------------------------------------------------
+  # Define the variables to be used to test in this object
+  #---------------------------------------------------------
+  before do
+    @version_file = 'test/support_files/test_version_grabber_file.md'
   end
 
-  def test_it_does_something_useful
-    assert true  #make this dummy test work
+  describe 'Test module Marskal' do
+
+    it 'Has a version_number' do
+      Marskal::VERSION.wont_be_nil
+    end
+
   end
+
+  describe 'Test module Marskal::Configuration' do
+
+    it 'Resets properly' do
+      skip 'test the configuration build and reset options'
+    end
+
+  end
+
+
 end